requirements-dev.txt
-r requirements.txt
black==19.10b0
coverage==5.0.3
flake8==3.7.9
isort==4.3.21
mypy==0.761
pylint==2.4.4
pytest-django==3.8.0
pytest==5.3.5
